+++ src/rust/src/lib.rs
@@ -155,7 +155,7 @@ mod _rust {
     #[pyo3::pymodule(gil_used = false)]
     mod declarative_asn1 {
         #[pymodule_export]
-        use crate::declarative_asn1::asn1::{decode_der, encode_der};
+        use crate::declarative_asn1::asn1::{decode_der, decode_der_annotated, encode_der};
         #[pymodule_export]
         use crate::declarative_asn1::types::{
             non_root_python_to_rust, AnnotatedType, Annotation, BitString, Encoding,